Python version compatibility fix for #628 (#630)
- #628

#628 works for Python 3.7, but not for higher versions of Python. It
seems that in higher versions of Python, `get_type_hints` will try to
resolve `ForwardRef`s and cause NameErrors.

The fix is to drop `get_type_hints` and use `object.__annotations__`
directly for the type hints. For the annotations to appear there
untouched, automatic string annotations (`from __future__ import
annotations`) needs to be removed. To remove automatic string
annotations, forward reference annotations need to be manually
stringified.

This PR does all of the above. It also passes all tests in Python 3.7
and Python 3.9.

Finally, we introduce a subclass of `TypeError` called `APITypeError` to
highlight when the exception is being hit at the top level :)