Fix 100G regression from #600 (#631)
TPCH 100G benchmark regressed after #600 , especially on Q4. Viewing the
task graph, it seems the issue is that tasks are getting run serially.

The issue is that we now request metadata within the physical plan. So
what can happen is:

1. Runner dispatches a task X.
2. Runner asks physical plan for the next task. The next task depends on
the output of task X.
3. Previously, this would get pipelined, and the runner can continue
dispatching tasks. But now, physical plan asks for the metadata of the
previous task, and thus blocks on X completing. This prevents new tasks
from being dispatched while X is completing; in the worst case, if X is
the only task executing, parallelism is completely starved.

The solution in this PR is to do batch dispatch. Batch dispatch is
already implemented, it just had to be turned on in default.

Also, this PR makes the memory requests more conservative (just 1x,
which is the amount the task is guaranteed to use).