Ignore multiple calls to set_runner_ray with kwarg noop_if_initialized (#728)
* Currently, if users call `set_runner_ray(...)` more than once per
process, they will incur a RuntimeError. The intention here is to be
very explicit that this behavior is not encouraged, as setting the
runner more than once can potentially corrupt the state of Daft's
caching.
* This PR adds a kwarg `noop_if_initialized`. If set to `True`,
subsequent calls to `set_runner_ray` are ignored and have no effect. We
log a warning when this happens.

Co-authored-by: Jay Chia <jaychia94@gmail.com@users.noreply.github.com>