Add optimization functions to expressions (#738)
* Implements expression methods that are used during query optimization
* It's a little annoying that we have to define logic to reconstruct
every expression every single time. One possible optimization that can
be made here is if we have each expression define a `.to_children`
method which will return an ordered Vec of its children, and a
`.from_children` which takes as input an ordered Vec of modified
children. But that optimization is left as an exercise for another PR.
* Testing is performed in `tests/optimizer`, but does not run on `main`
because `expressions2` is not yet integrated with main. All optimizer
tests pass on our Rust integration branch.

Co-authored-by: Jay Chia <jaychia94@gmail.com@users.noreply.github.com>