[RUST] Expression typing tests (#744)
# Summary

For every kernel in our codebase, test it against every permutation of
DataTypes in our codebase.

* Add logic per-kernel to determine which DataTypes it is valid for
* If this kernel should be a valid operation on these DataTypes:
    1. Run the kernel directly on the underlying `Series` objects
2. Assert that the returned `Series` has the same datatype as that
returned during schema resolution
* If the kernel is an invalid operation on these DataTypes:
    1. Assert that schema resolution fails as expected

## TODOs

1. Some TODOs had to be added into the code due to our incomplete
if/else implementation. I will address that in a separate PR so that our
if/else kernel will work on all the physical types.
3. We have pretty poor handling for `DataType.binary()` right now, and
many tests are failing with panics as we don't implement this in many
places. I left a TODO here to address in a separate PR since this is not
crucial for TPC-H at the moment.
4. Similarly, `DataType.date()` is poorly handled and is commented out
here

---------

Co-authored-by: Jay Chia <jaychia94@gmail.com@users.noreply.github.com>