Rust Series: Add strict arrow and explicit PyObject to creation methods. (#805)
Previously, when from_pylist was called on a non-Arrow-convertible type,
it would store the data as PyObjects (this being the primary codepath
for pyobject ingestion). This PR adds the option to explicitly request
arrow objects and error out if they are not arrow coercible, to allow
users to guard against accidentally storing PyObjects when they don't
mean to. Similarly, this PR adds the option to bypass the Arrow
conversion and explicitly store Python objects if they know for sure
that that's what they want.

- `Series.from_pylist` now has an additional optional parameter `pyobj`
defaulting to `allow` and can also be `disallow` or `force`.

---------

Co-authored-by: Xiayue Charles Lin <charles@eventualcomputing.com>