Get count working for PyObjects (#827)
For now, "count" will be the only aggregate we can implement on opaque
types like PyObject. Semantics are the same as Arrow - it counts
non-nulls, with Python None being a null.

Our count implementation already did not directly reference arrow2 data
type, but it calls arrow2 array null_count, which asks for the data type
under the hood (to check if it's NullType). So this PR reimplements it
on PseudoArrowArray and removes the check.

Also added a test.

---------

Co-authored-by: Xiayue Charles Lin <charles@eventualcomputing.com>