[RUST] Explode List types (#824)
* Implements explode on the Table level
(`table.explode([col("x")._explode()])`)
* Implements explode on the DataFrame level (`df.explode(col("x"))`)

**Discrepancy in API between the DataFrame and the Table:**

Note that (similar to aggregations) `Expression._explode()` is not
intended to be exposed to users, since running a projection on exploded
columns is a weird operation. Thus to a user, `explode` is a
dataframe-level operation, and we apply `._explode()` to each expression
in Python-land for schema resolution. However, since It's tricky to get
the "child expression" of an Explode expression in Python, when we call
this expression on the Table we will call it as Explode expressions as
well.

---------

Co-authored-by: Jay Chia <jaychia94@gmail.com@users.noreply.github.com>