PyObjects in Expressions: Cast Python -> Arrow (#853)
- Implement casting from Python to Arrow. The semantics is that first,
the appropriate Python type function is called; then, the Arrow cast is
called.
- Example: to cast some series [pyobj1, ...] to Int32, first Daft will
call `int(pyobj1)` in Python (which will delegate to `pyobj1.__int__` if
implemented). Then Daft puts the resulting list is put into a Series,
asserts that this is now an Arrow series, and casts it again to Int32
(this time following the existing arrow cast codepath).
- This PR makes a macro to do this, `pycast_then_arrowcast`, which takes
in the series, the Python cast, and the Arrow cast. The macro is used to
fill out the cast typematch for numerics, bool, binary, utf8. Null cast
returns Null, following existing Arrow behaviour. Date, etc casts are
left unimplemented because the Python conversion is not standardized.
- Tests included.
- Driveby: Update PyResult monad chains to use `?` operator instead of
`and_then`. This is only possible if you explicitly annotate the return
type of the py closure.

---------

Co-authored-by: Xiayue Charles Lin <charles@eventualcomputing.com>