Thread user-provided schema through to DataFrame reads (#1024)
* Adds a new `schema_hints` kwarg in our `read_*` APIs. This allows
users to specify up-front what the schema of their data will be

When provided, the `schema_hints` kwarg will disable all schema
inference. The "inferred" schema will be the one that was provided by
the user. Daft will cast all incoming data it is reading into the
provided schema.

**NOTE: The `read_csv(column_names=...)` kwarg had to be deprecated!**
Having both the option to pass in the schema and an option to just pass
in column names made it very difficult to reason about the behavior of
the API. If a user really needs this behavior (pass in the column names,
but infer the types), they can still perform a normal
`read_csv(has_header=False)` and then a `df.select(*[c_old.alias(c_new)
for c_old, c_new in zip(df.column_names, actual_column_names)])`. A
little niche, but at least there's a viable workaround.

---------

Co-authored-by: Jay Chia <jaychia94@gmail.com@users.noreply.github.com>