[FEAT] Locally unique semantic IDs for Expressions (#1243)
This PR contains code for generation of locally-unique (Table scope)
semantic IDs for Expressions. To get this ID for an expression, call
`expr.semantic_id(schema)`.

The semantics for these local IDs are that two identical expression
trees have the same ID. In particular,

- UDFs have consistent (not random) IDs. This is required for
Table-scope expression caching, and opposite of the Dataframe-level
requirement that distinct *executions* have distinct IDs.
- Globally-unique semantic IDs are not in this PR. They require a
dataframe identifier, which is not appropriate to include in the table
scope. Similarly, they are only used by the query planner, and
operations on individual Tables should not know about them.

This is intended to be used for common subexpression caching at the
Table evaluation level, as well as a building block for global field IDs
in the query planner.

---------

Co-authored-by: Xiayue Charles Lin <charles@eventualcomputing.com>