[FEAT] Add all projection optimization rules to new query planner. (#1288)
This PR includes all the rules we currently have for optimizing
Projections, but is organized slightly differently.

When run on a Projection node, ProjectionPushdown will:

1. Delete the node if it is unnecessary; or else
2. Trim columns in upstream Projections, Aggregates, and Source nodes;
or else
3. Create new projections behind other upstream nodes if possible. 

Finally, if a change has been made, this rule recurses on the new node
immediately. This ensures that old projection nodes that have become
newly redundant are pruned immediately, while still having a clean
separation of logic such that none of the logic in {creating, deleting,
modifying} nodes have to worry about the other.

----

There is also a rule for Aggregate (since it also projects columns from
its parent). If the Aggregate implicitly drops columns from its parent,
a Projection is created upstream to make the drop explicit. This rule
will only ever fire once, since projection pushdown monotonically
decreases the number of columns across the plan. Then, the newly created
projection will be optimized just like any other projection.

----

This PR also updates the tabular scan shim to include the ability to
pass in specific columns to read. We also rename schemas in source nodes
and associated structs to make it explicit whether the schema is of the
source or of the intended plan output.

---------

Co-authored-by: Xiayue Charles Lin <charles@eventualcomputing.com>
Co-authored-by: Clark Zinzow <clarkzinzow@gmail.com>