[FEAT] [New Query Planner] Refactor file globbing logic by exposing `FileInfos` to Python (#1307)
This PR refactors the file globbing logic by exposing `FileInfos` from
the new query planner to Python and using it for both the old query
planner and the new query planner. Hopefully the upcoming Rust-native
file globber will be able to leverage the `FileInfos` struct or
something akin to it, allowing it to be a ~drop-in replacement for the
existing Python file globber.