[CHORE] Fix List/FixedSizeList DataType to hold a dtype instead of Field (#1351)
Closes: #1264 
Closes: #994 

Previously, our `DataType::List/FixedSizeList` would hold a child Field
instead of DataType. This was causing a few problems:

1. When comparing equality of two schemas, we often hit weird cases
where the naming of the subfield was wrong
2. When creating a new ListArray, sometimes we could forget to correctly
name the child array (e.g. when using a growable) which would result in
pretty nasty failures

This PR:
1. Hardcodes the child Series' name to `"list"`
2. Changes List/FixedSizeList DataType to hold a dtype instead of Field,
which essentially removes the name from the type and makes it
"anonymous"

Note that this is consistent with PyArrow behavior as well - in more
recent versions of PyArrow, they have started to ignore equality of the
subfield's name.

---------

Co-authored-by: Jay Chia <jaychia94@gmail.com@users.noreply.github.com>