[BUG] Fix public API decorator type annotations. (#1397)
This PR fixes the type annotations for functions/methods that are
wrapped with the public API decorator. Namely, we replace `Callable[...,
Any]` for the input and return type of the decorator with `Callable[P,
T]`, where `T = TypeVar("T")` and `P = ParamSpec("P")`; we then bind `T`
to the return type and `P.args` and `P.kwargs` to the args and kwargs,
respectively, of the wrapper function. This ensures that the type
annotations of the wrapped function/method are properly propagated to
the returned wrapper function.