[BUG] Fix num input partitions in coalesce. (#1442)
This PR fixes the number of input partitions in our `Coalesce`
implementation. The new query planner was incorrectly using the current
logical node's `num_partitions` for the number of _input_ partitions,
which for a `Coalesce` op, is equal to the _output_ number of
partitions; we should instead be using the input (child) logical node's
`num_partitions`.

The microbenchmarking failure also illustrated that we didn't have any
unit testing coverage of `Coalesce` (i.e. `df.into_partitions(n)` where
`n` is smaller than the current number of partitions), so we also add a
sanity unit test.

Closes #1434