[FEAT] [New Query Planner] [2/N] Push partition spec into physical plan, remove Coalesce logical op. (#1540)
This PR pushes the partition spec concept into the physical plan; this
is required in order to defer partition information gathering (e.g.
globbing) until plan translation, and also helps unify some past
duplication of partitioning-based logic between the logical plan and
logical -> physical translation.

In addition to push all partition spec logic to plan translation and the
physical plan, the following major changes were made to facilitate this:
- The `Coalesce` logical op is absorbed into the `Repartition` logical
op; the switching between a `Split` or `Coalesce` is now made at logical
-> physical translation time.
- `df.repartition()` now takes an optional `num_partitions` arg, and
`num_partitions` is now optional for the logical `Repartition` op; this
is needed to support the "hash repartition by these columns to the same
number of partitions as the input" use case, e.g. for writing out
partitioned CSV/JSON datasets.
- Support for `df.num_partitions()` has been preserved, by triggering
optimization + plan translation to fulfill the query; we can look at
caching this in the future, so we don't have to redo globbing once we've
moved globbing to plan translation time. ~`df.num_partitions()` API is
removed, since this can't be determined based on just the logical plan
alone. We could add this back and trigger plan optimization +
translation to get this answer, although this will end up triggering
nontrivial computation for the globbing backend, so we may want to (1)
only trigger enough execution in order to get the `num_partitions`
answer, and (2) cache that execution.~


## TODOs Before Merging

- [x] Add `DropRepartitions` optimizations as a split between a logical
optimization rule (for non-partition spec logic) and a plan
translation-time optimization (for partition spec logic).
- [x] Add old logical `Project` unit tests around partition spec munging
to physical `Project` operator

---------

Co-authored-by: Jay Chia <17691182+jaychia@users.noreply.github.com>