[BUG] Use schema_hints as hints instead of definitive schema (#1636)
Addresses #1599 

Instead of using `schema_hints` as a definitive schema, use them as
'hints' as to the intended datatype of each column.
This is implemented via running schema inference first, then applying
the 'hints' onto the inferred schema.

Tests:
- Added tests for for read_csv, read_json, read_parquet

Feedback greatly appreciated! Let me know if this is the correctly
intended behaviour, and also if the code can be optimized/refactored
since this is my first time writing Rust!