[FEAT] [Join Optimizations] Add broadcast join. (#1706)
This PR adds a broadcast join implementation as a new join strategy,
where all partitions of a small table are broadcasted to each partition
in the larger table, such that we do a local (hash) join of the entire
small table with each individual partition of the larger table.

## Query Planning

The query planner chooses the broadcast join as its join strategy if one
of the sides of the join is smaller than a preconfigured broadcasting
threshold (set to 10 MiB by default, but is user-configurable).

**Note:** If the smaller side of the join is the right side, we invert
the join for planning and scheduling simplicity so we can always
broadcast the left side; we then swap back to the correct join ordering
when performing the local joins. This means that we always form the
probe table on the left side of the join; a future optimization
(applicable to both the broadcast join and the hash join) would be to
have local joins build the probe table on the smaller side while
preserving the expected column ordering. We would still need to always
build the probe table on the left side of the join if we need to
preserve the row-ordering of the right side of the join, e.g. if the
right side of the join is range-partitioned AND we're doing a broadcast
join.

## Query Scheduling

All partitions for the broadcasting side of the join are first
materialized. Then, as each partition on the receiving side of the join
materialize, we dispatch a hash join task joining all broadcaster
partitions with that single receiving-side partition.

## TODOs

- [x] Test coverage.
- [ ] (Follow-up?) TPC-H benchmarking demonstrating speedup due to use
of broadcast join.
- [ ] (Follow-up) In local joins, build the probe table on the smaller
side of the join.
- [ ] (Follow-up) Add table size approximations for operators that
affect cardinality.