[FEAT] Add getter for Struct and List expressions (#1775)
Adds getters for nested and struct types, which enables the
access of elements in `ListArray`, `FixedSizeListArray`, and
`StructArray` types.

For lists:
```
Expression.list.get(idx, default=None)
``` 
This method takes in an index, which can be a single value or an
expression, and optionally a default value, which will be set when the
index is out of bounds. It retrieves the value at index of each list in
the array.


For structs:
```
Expression.struct.get(name)
``` 
This method takes in the name of the field to get, and returns an
expression of that struct's field.