[BUG] [Parquet] Fix double-await on `JoinHandle`s concurrency bug in Parquet reader. (#1841)
This PR fixes a double-await concurrency bug on `JoinHandle`s in the
Parquet reader. The bug appears to arise when an a byte range fetch
fails: we currently short-circuit that fetch and return an error to the
stream mapper without changing the range cache state. This means that
future fetches that overlap with that byte range will try to `.await` on
the same `JoinHandle`, which is not allowed.

This PR fixes this by storing a `Result<Bytes, Error>` in
`RangeCacheState::Ready()`, so we can stash errors from past fetches.

## TODOs

- [x] Validate on TPC-H Q1 benchmark reproduction.
- [ ] Ensure that PR matches at least the status quo in terms of
fault-tolerance