[FEAT] Enable Requester Pay for S3 reads (#1856)
This PR enables reads from S3 buckets with a requester pays policy. This
configuration is set in the S3 IO Config.

Todo: integration tests? will we need to set up a requester pays bucket
for this? I did test it on the 1tr row parquets tho, so it is working.