[FEAT] Top level global expressions (#2000)
This PR allows the usage of aggregation expressions (+alias) at the top
level of an agg. For example:
- `df.agg(col("x").mean().alias("x_mean"))`
- `df.groupby(col("a")).agg((col("a")+col("b")).max())`

In addition, it also sets up the ability to pass in expressions into
dataframe operations as either individual arguments `df.agg(expr1,
expr2)` or in a list `df.agg([expr1, expr2])`. This PR implements this
for all the aggregation functions, and a future PR can use the defined
`ColumnInputOrListType` type and `DataFrame._inputs_to_expressions`
function to easily implement it for any operation.

With this PR, the previous syntax for aggregations, which uses a list of
tuples to express the columns to aggregate, is still supported but
raises a deprecation warning.