[FEAT] Add str.extract_all expression (#2038)
`extract_all(self, pattern: str | Expression, index: int = 0)` 
Extracts the specified match group from all regex matches in each string
in a string column. If index is 0, the entire match is returned. If the
pattern does not match or the group does not exist, a null value is
returned.

As a drive by: 
- Refactored + made some optimizations to the existing str.extract
functionality