[CHORE] Add tpch test for read sql (#2026)
This PR adds tpch tests via read_sql. Our tpch integration test setup
already creates a sqlite table, which makes it easy to run `read_sql`
against for complex query validation.

As a drive by in addition to adding tpch tests, this PR gates the Date
and Timestamp literal to sql translations. This is because the semantics
for date and timestamp literals were not as general as expected.