[FEAT] [New Executor] [1/N] Move physical plan scheduler to new crate, misc. refactorings + drive-bys for new executor (#2339)
This PR moves the physical plan scheduler to a new `daft-scheduling`
crate in order to avoid cycle dependencies for the new executor, which
depends on `daft-plan`. New dependency tree will be:
```
daft-plan   <--  daft-execution
       ^            ^
        \          /
      daft-scheduling 
```

This PR also contains misc. refactorings and drive-bys for follow-up new
executor PRs, and will be the bottom of a stack of PRs.