[DOCS] the project currently uses ray which is not compatible with python 3.12 (#2354)
I was setting up daft dev environment on my local and was following
along the contribution guidelines. my default python env is currently at
`3.12` and noticed that the `make .venv` keeps failing due to dependency
on `ray[data]==2.10.0`. the failure happens because the current
`ray-data` is only compatible till `python3.11`.