[FEAT] Add manual auth for GCS and Iceberg GCS auth support (#2393)
Tabular credentials vending should now work for Google Cloud Storage.
Additionally, now users can manually pass in a credentials file or
string instead of relying on it to be picked up from the environment.

Locally tested with both a credentials file and OAuth2 access token.
Working on testing with Tabular as well. I'm unsure how to properly
write tests for it though, or if it's worth doing.