[FEAT] Add struct get syntactic sugar (#2367)
Adds the ability to query struct and map fields by using the dot syntax,
such as `col("a.b")` turning into `col("a").struct.get("b")`.

This PR also includes a minor refactor of agg expression checking and
extraction, just moving it out of the builder and into the
`resolve_expr` and `resolve_aggexpr` functions that also deal with the
syntactic sugar. I changed this since we were talking about how brittle
it would be to do the syntactic sugar conversion in the builder and
realized that it applies to aggregations too.