[FEAT] Add MinHash expression (#2431)
# MinHash Algorithm

This PR adds an expression for computing the MinHash of a series of
strings.

## What is MinHash?

MinHash is an algorithm that computes a special collection of hashes for
a set. The algorithm is simple: for every element in the set, compute
its hash, and take the minimum of the hashes. Then, repeat using several
different hash functions. These minhashes have the nice property that,
when computed on two sets, the proportion of hashes that are equal is an
unbiased estimator of the Jaccard similarity of the sets, meaning that
it approximates how often elements from the sets overlap. We apply this
to strings by computing the hashes on ngrams of the space-separated
words in the string.

One optimization that can be performed is by permuting one initial hash
instead of using several different hashes. In our implementation, this
involves taking a random linear function and applying it to our current
hash, modulo a large prime. Repeating this several times gives us a
pairwise independent hash family, with the benefit of being very easy to
compute. One could use higher-order polynomials, which gives us better
guarantees on correlation between hashes, but comes at the cost of being
harder to compute.

## Computing the MinHash

### Naive algorithm

The naive algorithm for computing the MinHash in a DataFrame is to
compute the ngrams, compute all their hashes, then afterwards use
grouping to take the minimum over the hashes. This is very undesirable
on larger datasets because grouping is hard to parallelize and the
number of ngrams can be quite large.

### UDF version

We can compute the MinHash of a string more efficiently using a UDF.
Given a string as input, we can iterate over the ngrams and maintain the
minimum hash as we go along. With this, we never need to keep more
hashes in memory than we need.

### Rust implementation

This PR allows us to replace this UDF with a built-in expression in
Rust, along with several optimizations. We avoid string copying as much
as possible, instead relying on a sliding window of string slices to
calculate hashes. We also use a more efficient initial hash algorithm,
MurmurHash, which gives us increased efficiency with the tradeoff of
reduced cryptographic security, which we don't really need for our
purposes. However, users should be aware that these minhashes should not
be used on sensitive data.

In more detail, for each string, we do the following:
- Find all the spaces in the string to separate the tokens,
- Iterate over the ngrams in the string, and for each ngram:
- Compute the initial hash using MurmurHash,
- Calculate `num_hashes` permutations of this hash,
- If a permutation is less than the current minimum, set it as the new
minimum.

We also use SIMD to optimize the calculations of the permutations.

## Results

> Note: these tests were performed before the SIMD optimizations. Also,
they use a lower than usual hash count--128 is a much more realistic
number.

Shown below is the flame graph of a run of the UDF MinHash algorithm.
This was executed on 35 MB of data from the [C4
dataset](https://huggingface.co/datasets/allenai/c4), which contains the
the text and URL of many webpages collected using a web crawler. We use
the uncleaned version of the dataset, since the cleaned version has
already had duplicates removed. As the whole dataset is over 2 TB in
size, we take a sample of 10k rows of the data by streaming it in.

We run the minhash algorithm with parameters `num_hashes = 8` and
`ngram_size = 8`. The UDF takes a total of 75.83 seconds to run over all
10k rows.


![results_10k_noexpr](https://github.com/Eventual-Inc/Daft/assets/32619800/0760bd35-93fe-45fd-a76d-4eadfcbe22b9)

Here are the results with using the `minhash` expression added in this
PR, using the same parameters. The expression takes a total of 0.46
seconds to run over all 10k rows, an increase in speed of around 165
times.


![results_10k_expr](https://github.com/Eventual-Inc/Daft/assets/32619800/46f8aa33-7f62-4464-b9f2-3cfe0c697bec)
 
For a more large-scale test, we run the `minhash` expression on 1
million rows of the dataset, totalling 3.35 GB. For this flame graph, we
ran all steps of deduplicating the dataset, and the MinHash step took
around 130.19 seconds.
 

![results_1m_full](https://github.com/Eventual-Inc/Daft/assets/32619800/1953fec1-84e6-4705-b7d4-b721f23fcedd)

It's worth noting that this algorithm is parallelizable, as rows do not
depend on each other. Thus, the runtime should scale inversely with the
number of cores used.

## Future work

There are a few more optimizations that could be made to this algorithm.
Here are a couple of examples:
1. Using a rolling hash: a rolling hash is very suitable for this
algorithm, as we are calculating hashes on a sliding window of the
original string. As hashing currently takes around 18% of the
algorithm's runtime, this could result in quite an increase in
efficiency.
2. Better vectorization: Instead of computing all permutations one by
one for every ngram, we can compute them for all ngrams at the same
time. This has potential to be more efficient and cache-friendly, though
it would need to be benchmarked.
3. Faster arithmetic: Using 32-bit operations to calculate hash
permutations instead of 64-bit could result in up to a 2x speedup.

We can also extend the algorithm to run on hashes of other objects, such
as images or videos, which would allow them to be deduplicated as well.

### Sources
- https://huggingface.co/blog/dedup
- https://en.wikipedia.org/wiki/MinHash
- https://en.wikipedia.org/wiki/K-independent_hashing