[FEAT] List slice expression (#2479)
Adds an expression that returns a specified subset of elements from a
list.

The API to use this expression is `daft.Expression.list.slice(start: int
| Expression, end: int | Expression)`.

- `start` is the 0-indexed position of the list to start retrieving
elements. If this value is negative, then it's the position from the
_end_ of the array. i.e. `-1` points to the last element of the array
- `end` is the 0-indexed position of the list to stop retrieving
elements. If this value is negative, then it's the position from the
_end_ of the array. i.e. `-1` points to the last element of the array

---------

Co-authored-by: Desmond Cheong <desmond@Desmonds-MBP.localdomain>