[BUG] Fix anti-join on different column names (#2477)
Closes #2475.

A bit of a bandaid fix. The issue was, while selecting the join
strategy, for checking whether sort-merge join works, the joined columns
need to be primitive types. However, in the plan we only have access to
the output schema (as far as I know, maybe we can extract it from the
children?). In a semi- or anti-join, the right column does not end up in
the output schema, so the primitive type check panicked while looking
for it. This issue does not appear if the left column has the same name
as the right.

My fix is that, since we only support SMJ for inner joins, I move the
join type check before the primitive check. This will need to be fixed
later if we support SMJ for anti-joins, so I added a comment.