[FEAT]: dyn function registry (#2466)
this adds a new variant to `Expr`. `Expr::ScalarFunction` which is
pretty similar to `FunctionExpr`, but uses dynamic dispatch and a
registry instead of the enum variants.

The registry is inspired by datafusion's function registry. 

when an expr is serialized, it just serializes the name and the inputs.

for example
`col("text").hash(seed=42)`

```js
{
  "name": "hash",
  "inputs": [
    `col('text')`, // serialized repr of this
    `lit(42)` // serialized repr of this
   ]
}
```

then when deserializing, it just fetches the appropriate function from
the registry. _(errorring if no matches found)_.



Also just to make sure everything works, I refactored the `hash`
function to use the new paradigm.