[FEAT] List chunk expression (#2491)
Adds an expression that splits lists into fixed sized lists with chunks
of a given size. If the number of elements is not a multiple of the
given chunk size, elements that do not fit into a chunk are discarded.

The API to use this expression is `daft.Expression.list.slice(size:
int)`, where `size` is the size of chunks to use. The given size must be
an integer greater than 0.

For example, a list [1, 2, 3, 4, 5] with a chunk size of 2 gives [[1,
2], [3, 4]].

---------

Co-authored-by: Desmond Cheong <desmond@Desmonds-MacBook-Pro.local>